requirements.txt
# Core GUI Framework
PyQt5>=5.15.0

# Gemini API Client
google-generativeai>=0.3.0

# Voice Processing
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Scientific Computing
numpy>=1.21.0

# API & Web
requests>=2.25.0

# System Monitoring
psutil>=5.8.0

# Environment Variables
python-dotenv>=0.19.0

# Optional: Local VLM
# torch>=2.0.0
# transformers>=4.35.0
# accelerate>=0.20.0